requirements.txt
ollama
openai-whisper
torch
torchaudio
soundfile
edge-tts
gradio
pytz
sentence-transformers
faiss-cpu
langchain-text-splitters
pypdf2
numpy
accelerate
scipy
